+++ src/Sentry.DiagnosticSource/Internal/DiagnosticSource/SentrySqlListener.cs
@@ -11,9 +11,6 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         Execution
     };
 
-    internal const string ConnectionExtraKey = "db.connection_id";
-    internal const string OperationExtraKey = "db.operation_id";
-
     internal const string SqlDataWriteConnectionOpenBeforeCommand = "System.Data.SqlClient.WriteConnectionOpenBefore";
     internal const string SqlMicrosoftWriteConnectionOpenBeforeCommand = "Microsoft.Data.SqlClient.WriteConnectionOpenBefore";
 
@@ -41,31 +38,32 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         _options = options;
     }
 
+    private static void SetDatabaseName(ISpan span, string databaseName)
+    {
+        Debug.Assert(databaseName != string.Empty);
+
+        span.SetExtra(OTelKeys.DbName, databaseName);
+    }
+
     private static void SetConnectionId(ISpan span, Guid? connectionId)
     {
         Debug.Assert(connectionId != Guid.Empty);
 
-        span.SetExtra(ConnectionExtraKey, connectionId);
+        span.SetExtra(SqlKeys.DbConnectionId, connectionId);
     }
 
     private static void SetOperationId(ISpan span, Guid? operationId)
     {
         Debug.Assert(operationId != Guid.Empty);
 
-        span.SetExtra(OperationExtraKey, operationId);
+        span.SetExtra(SqlKeys.DbOperationId, operationId);
     }
 
-    private static Guid? TryGetOperationId(ISpan span) =>
-        span.Extra.TryGetValue(OperationExtraKey, out var key) && key is Guid guid
-            ? guid
-            : null;
+    private static Guid? TryGetOperationId(ISpan span) => span.Extra.TryGetValue<string, Guid?>(SqlKeys.DbOperationId);
 
-    private static Guid? TryGetConnectionId(ISpan span) =>
-        span.Extra.TryGetValue(ConnectionExtraKey, out var key) && key is Guid guid
-            ? guid
-            : null;
+    private static Guid? TryGetConnectionId(ISpan span) => span.Extra.TryGetValue<string, Guid?>(SqlKeys.DbConnectionId);
 
-    private void AddSpan(SentrySqlSpanType type, string operation, object? value)
+    private void AddSpan(string operation, object? value)
     {
         var transaction = _hub.GetTransactionIfSampled();
         if (transaction == null)
@@ -74,25 +72,11 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         }
 
         var operationId = value?.GetGuidProperty("OperationId");
-
-        switch (type)
-        {
-            case SentrySqlSpanType.Connection when transaction.GetDbParentSpan().StartChild(operation) is
-                { } connectionSpan:
-                SetOperationId(connectionSpan, operationId);
-                break;
-
-            case SentrySqlSpanType.Execution when value?.GetGuidProperty("ConnectionId") is { } connectionId:
-                var parent = TryGetConnectionSpan(transaction, connectionId) ?? transaction.GetDbParentSpan();
-                var span = TryStartChild(parent, operation, null);
-                if (span != null)
-                {
-                    SetOperationId(span, operationId);
-                    SetConnectionId(span, connectionId);
-                }
-
-                break;
-        }
+        var connectionId = value?.GetGuidProperty("ConnectionId");
+        var parent = transaction.GetDbParentSpan();
+        var span = parent.StartChild(operation);
+        SetOperationId(span, operationId);
+        SetConnectionId(span, connectionId);
     }
 
     private ISpan? GetSpan(SentrySqlSpanType type, KeyValuePair<string, object?> kvp)
@@ -107,24 +91,15 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         {
             case SentrySqlSpanType.Execution:
                 var operationId = kvp.Value?.GetGuidProperty("OperationId");
-                if (TryGetQuerySpan(transaction, operationId) is not { } querySpan)
+                if (TryGetQuerySpan(transaction, operationId) is { } querySpan)
                 {
-                    _options.LogWarning(
-                        "Trying to get an execution span with operation id {0}, but it was not found.",
-                        operationId);
-                    return null;
+                    return querySpan;
                 }
 
-                if (querySpan.ParentSpanId == transaction.SpanId &&
-                    TryGetConnectionId(querySpan) is { } spanConnectionId &&
-                    querySpan is SpanTracer executionTracer &&
-                    TryGetConnectionSpan(transaction, spanConnectionId) is { } spanConnectionRef)
-                {
-                    // Connection Span exist but wasn't set as the parent of the current Span.
-                    executionTracer.ParentSpanId = spanConnectionRef.SpanId;
-                }
-
-                return querySpan;
+                _options.LogWarning(
+                    "Trying to get an execution span with operation id {0}, but it was not found.",
+                    operationId);
+                return null;
 
             case SentrySqlSpanType.Connection:
                 var connectionId = kvp.Value?.GetGuidProperty("ConnectionId");
@@ -142,9 +117,6 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
                 throw new ArgumentOutOfRangeException(nameof(type), type, null);
         }
     }
-
-    private static ISpan? TryStartChild(ISpan? parent, string operation, string? description) =>
-        parent?.StartChild(operation, description);
 
     private static ISpan? TryGetConnectionSpan(ITransaction transaction, Guid? connectionId) =>
         connectionId == null
@@ -173,8 +145,7 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         }
 
         var operationId = value.GetGuidProperty("OperationId");
-        var connectionId = value.GetGuidProperty("ConnectionId");
-        if (operationId == null || connectionId == null)
+        if (operationId == null)
         {
             return;
         }
@@ -182,7 +153,16 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
         var spans = transaction.Spans.Where(span => span.Operation is "db.connection").ToList();
         if (spans.Find(span => !span.IsFinished && TryGetOperationId(span) == operationId) is { } connectionSpan)
         {
-            SetConnectionId(connectionSpan, connectionId);
+            if (value.GetGuidProperty("ConnectionId") is { } connectionId)
+            {
+                SetConnectionId(connectionSpan, connectionId);
+            }
+
+            if (value.GetStringProperty("Connection.Database") is { } dbName)
+            {
+                connectionSpan.Description = dbName;
+                SetDatabaseName(connectionSpan, dbName);
+            }
         }
     }
 
@@ -198,22 +178,22 @@ internal class SentrySqlListener : IObserver<KeyValuePair<string, object?>>
             {
                 // Query
                 case SqlMicrosoftBeforeExecuteCommand or SqlDataBeforeExecuteCommand:
-                    AddSpan(SentrySqlSpanType.Execution, "db.query", kvp.Value);
+                    AddSpan("db.query", kvp.Value);
                     return;
                 case SqlMicrosoftAfterExecuteCommand or SqlDataAfterExecuteCommand
                     when GetSpan(SentrySqlSpanType.Execution, kvp) is { } commandSpan:
-                    commandSpan.Description = kvp.Value?.GetProperty("Command")?.GetStringProperty("CommandText");
+                    commandSpan.Description = kvp.Value?.GetStringProperty("Command.CommandText");
                     commandSpan.Finish(SpanStatus.Ok);
                     return;
                 case SqlMicrosoftWriteCommandError or SqlDataWriteCommandError
                     when GetSpan(SentrySqlSpanType.Execution, kvp) is { } errorSpan:
-                    errorSpan.Description = kvp.Value?.GetProperty("Command")?.GetStringProperty("CommandText");
+                    errorSpan.Description = kvp.Value?.GetStringProperty("Command.CommandText");
                     errorSpan.Finish(SpanStatus.InternalError);
                     return;
 
                 // Connection
                 case SqlMicrosoftWriteConnectionOpenBeforeCommand or SqlDataWriteConnectionOpenBeforeCommand:
-                    AddSpan(SentrySqlSpanType.Connection, "db.connection", kvp.Value);
+                    AddSpan("db.connection", kvp.Value);
                     return;
                 case SqlMicrosoftWriteConnectionOpenAfterCommand or SqlDataWriteConnectionOpenAfterCommand:
                     UpdateConnectionSpan(kvp.Value);